+++ examples/resources/sdwan_service_lan_vpn_interface_svi_feature_associate_dhcp_server_feature/import.sh
@@ -1,0 +1,1 @@
+terraform import sdwan_service_lan_vpn_interface_svi_feature_associate_dhcp_server_feature.example "f6b2c44c-693c-4763-b010-895aa3d236bd"